SW-5506 Make the 'variable manifest' CSV work with the latest CSV format
